requirements.txt
Flask==2.3.2
openai==1.30.0
python-dotenv==1.0.0
gtts==2.3.2
